requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
dj-database-url>=2.1,<3.0
psycopg2-binary>=2.9,<3.0
gunicorn>=21.0,<23.0
whitenoise>=6.6,<8.0
python-dotenv>=1.0,<2.0
Pillow>=10.0,<12.0
django-filter>=23.5,<25.0
django-extensions>=3.2,<4.0
django-jazzmin>=3.0,<4.0
weasyprint>=60.0,<70.0
pytest>=7.4,<9.0
pytest-django>=4.5,<5.0
pytest-cov>=4.1,<6.0
factory-boy>=3.3,<4.0